Simplify test resolution of Vite dependency
