+++ packages/react-router-dev/vite/build.ts
@@ -17,7 +17,7 @@ import {
 } from "../config/config";
 import type { RouteManifestEntry, RouteManifest } from "../config/routes";
 import invariant from "../invariant";
-import { preloadViteEsm } from "./import-vite-esm-sync";
+import { preloadVite, getVite } from "./vite";
 
 function getAddressableRoutes(routes: RouteManifest): RouteManifestEntry[] {
   let nonAddressableIds = new Set<string>();
@@ -235,8 +235,8 @@ export async function build(
   }: ViteBuildOptions
 ) {
   // Ensure Vite's ESM build is preloaded at the start of the process
-  // so it can be accessed synchronously via `importViteEsmSync`
-  await preloadViteEsm();
+  // so it can be accessed synchronously via `getVite`
+  await preloadVite();
 
   let viteConfig = await resolveViteConfig({ configFile, mode, root });
 
@@ -251,7 +251,7 @@ export async function build(
 
   let { reactRouterConfig } = ctx;
 
-  let vite = await import("vite");
+  let vite = getVite();
 
   async function viteBuild({
     ssr,
