+++ packages/react-router-dev/vite/dev.ts
@@ -1,7 +1,7 @@
 import type * as Vite from "vite";
 import colors from "picocolors";
 
-import { preloadViteEsm } from "./import-vite-esm-sync";
+import { preloadVite, getVite } from "./vite";
 import * as profiler from "./profiler";
 
 export interface ViteDevOptions {
@@ -34,10 +34,10 @@ export async function dev(
   }: ViteDevOptions
 ) {
   // Ensure Vite's ESM build is preloaded at the start of the process
-  // so it can be accessed synchronously via `importViteEsmSync`
-  await preloadViteEsm();
+  // so it can be accessed synchronously via `getVite`
+  await preloadVite();
+  let vite = getVite();
 
-  let vite = await import("vite");
   let server = await vite.createServer({
     root,
     mode,
