+++ packages/react-router-dev/vite/vite-node.ts
@@ -3,7 +3,7 @@ import { ViteNodeRunner } from "vite-node/client";
 import { installSourcemapsSupport } from "vite-node/source-map";
 import type * as Vite from "vite";
 
-import { importViteEsmSync, preloadViteEsm } from "./import-vite-esm-sync";
+import { preloadVite, getVite } from "./vite";
 
 export type Context = {
   devServer: Vite.ViteDevServer;
@@ -14,8 +14,8 @@ export type Context = {
 export async function createContext(
   viteConfig: Vite.InlineConfig = {}
 ): Promise<Context> {
-  await preloadViteEsm();
-  const vite = importViteEsmSync();
+  await preloadVite();
+  const vite = getVite();
 
   const devServer = await vite.createServer(
     vite.mergeConfig(
